basic isAD nFunction goes through compilation as nFunction or nClass method
