+++ nCompiler/NAMESPACE
@@ -64,8 +64,8 @@ export(nDeparse)
 export(nDerivs)
 export(nDim)
 export(nType)
+export(nCompile1)
 export(nCompile)
-export(nCompile2)
 export(writePackage2)
 export(nClass)
 export(nFunction)
